告警

+++ services/storage_daemon/mtp/src/mtp_device_monitor.cpp
@@ -331,8 +331,8 @@ int32_t MtpDeviceMonitor::HasMTPDevice(bool &hasMtp)
     }
     for (UsbDevice &dev : deviceList) {
         uint8_t deviceClass = static_cast<uint8_t>(dev.GetClass());
-        uint16_t idVendor = static_cast<uint8_t>(dev.GetVendorId());
-        uint16_t idProduct = static_cast<uint8_t>(dev.GetProductId());
+        uint16_t idVendor = static_cast<uint16_t>(dev.GetVendorId());
+        uint16_t idProduct = static_cast<uint16_t>(dev.GetProductId());
         LOGI("device class is %{public}d, vendor id is %{public}d, product id is %{public}d.",
              deviceClass, idVendor, idProduct);
         if (LIBMTP_check_is_mtp_device(deviceClass, idVendor, idProduct)) {
